requirements.txt
Flask
python-dotenv
SQLAlchemy
Flask-Cors
